requirements.txt
# BharatSecure Touchless HCI - Python Dependencies
# All open-source, zero cost

# Core Vision & ML
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.4
scikit-learn==1.3.2

# Deep Learning (TFLite for Jetson edge inference)
tflite-runtime==2.13.0        # Jetson-optimised; use tensorflow==2.13.0 on x86

# Security & Crypto
cryptography==41.0.7
pycryptodome==3.20.0

# Federated Learning utilities
torch==2.0.1                  # For federated weight aggregation (CPU/CUDA)
# torchvision==0.15.2         # Uncomment if needed

# Dashboard & API
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.35.1

# Data & Logging
pandas==2.1.4
sqlite3                        # stdlib, no install needed
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
colorlog==6.8.0
tqdm==4.66.1
requests==2.31.0
